drm/rockchip: vop: Calc dsp hold timeout use actual vrefresh
When crtc disables, we will wait for the dsp hold completion to ensure
that the current frame is fully displayed before disabling the VOP.

Now the dsp hold completion wait timeout is set to a constant value,
which may not apply to all scenarios.

This patch will calculate the corresponding wait time based on the
actual vrefresh, up to a maximum of 10 frames.

Change-Id: Ice60af969fb06ca949b53474f6c9817ba07f50fd
Signed-off-by: Chaoyi Chen <chaoyi.chen@rock-chips.com>